Changed ticker to use flat files rather than SQLite3, added update_tickerinfo.php for cron (rather than autoupdate via client)
